+++ src/main/java/gravit/code/notification/support/NotificationPushSender.java
@@ -1,0 +1,35 @@
+package gravit.code.notification.support;
+
+import gravit.code.fcm.dto.internal.PushMessage;
+import gravit.code.fcm.service.FcmService;
+import gravit.code.fcm.service.FcmTokenQueryService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+import java.util.Map;
+
+@Component
+@RequiredArgsConstructor
+public class NotificationPushSender {
+
+    private final FcmTokenQueryService fcmTokenQueryService;
+    private final FcmService fcmService;
+
+    public void pushToUser(
+            long userId,
+            Map<String, String> data,
+            String message
+    ) {
+        List<String> tokens = fcmTokenQueryService.getTokensByUserIds(List.of(userId))
+                .get(userId);
+
+        if (tokens == null || tokens.isEmpty()) {
+            return;
+        }
+
+        PushMessage pushMessage = PushMessage.of(tokens, message, null, data);
+
+        fcmService.sendNotifications(List.of(pushMessage));
+    }
+}
